image_comparision/requirements.txt
Pillow==9.4.0
sentence_transformers==2.2.2